apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
python-dotenv==1.0.1
sentence-transformers==3.0.0
numpy==1.26.4
scikit-learn==1.5.0
rank-bm25==0.2.2
groq==0.9.0
reportlab==4.2.0
python-multipart==0.0.9
aiofiles==23.2.1
tqdm==4.66.4
requests==2.32.3